% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genus_fcts}
\alias{genus_fcts}
\title{Food composition tables for GENuS}
\format{
A data frame with the following attributes::
\describe{
  \item{file}{Source file - from Harvard Dataverse website}
  \item{region}{Region}
  \item{genus_food_id}{GENuS food id}
  \item{genus_food_name}{GENuS food name}
  \item{fct_food_name}{FCT food name}
  \item{nutrient}{Nutrient}
  \item{units}{Nutrient units}
  \item{units_per_100g}{Nutrient per 100g}
}
}
\source{
Smith, M.R., Micha, R., Golden, C.D., Mozaffarian, D., Myers, S.S. (2016) Global Expanded Nutrient Supply (GENuS) model: a new method for estimating the global dietary supply of nutrients. PloS One 11(1): p.e0146976. https://doi.org/10.1371/journal.pone.0146976
}
\usage{
genus_fcts
}
\description{
Food composition tables organized by the commodities included in the Global Expanded Nutrient Supply model (GENuS) dataset. Their construction and use are described in the GENuS methods paper (citation below). Citations for original source tables and units for each nutrient are found in accompanying "readme" file.
}
\keyword{datasets}
