% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genus_food}
\alias{genus_food}
\title{Edible food supply by country and year (1961-2011)}
\format{
A data frame with the following attributes::
\describe{
  \item{iso3}{ISO3 standardized}
  \item{country}{Country standardized}
  \item{iso3_orig}{ISO3 in original GENuS data}
  \item{country_orig}{Country in original GENuS data}
  \item{year}{Year}
  \item{food}{Food}
  \item{g_person_day}{Supply (g/person/day)}
}
}
\source{
Smith, M.R., Micha, R., Golden, C.D., Mozaffarian, D., Myers, S.S. (2016) Global Expanded Nutrient Supply (GENuS) model: a new method for estimating the global dietary supply of nutrients. PloS One 11(1): p.e0146976. https://doi.org/10.1371/journal.pone.0146976
}
\usage{
genus_food
}
\description{
Edible food supplies for 225 individual food types in grams per person per day by country, separated by year (1961-2011).
}
\details{
Countries with data marked with a "*" have insufficient data to approximate their food supplies. For modern countries that were once part of another country (e.g., Armenia in USSR), older values reflect those of their predecessor.
}
\keyword{datasets}
