% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genus_food_2011}
\alias{genus_food_2011}
\title{Edible Food by Age and Sex (2011)}
\format{
A data frame with the following attributes:
\describe{
  \item{iso3}{ISO3 standardized}
  \item{country}{Country standardized}
  \item{iso3_orig}{ISO3 in original GENuS data}
  \item{country_orig}{Country in original GENuS data}
  \item{age_range}{Age range (yrs)}
  \item{sex}{Sex}
  \item{food}{Food}
  \item{units_long}{Units, long form}
  \item{units_short}{Units, short form}
  \item{supply_med}{Supply, median}
  \item{supply_lo}{Supply, 2.5th percentile}
  \item{supply_hi}{Supply, 97.5th percentile}
}
}
\source{
Smith, M.R., Micha, R., Golden, C.D., Mozaffarian, D., Myers, S.S. (2016) Global Expanded Nutrient Supply (GENuS) model: a new method for estimating the global dietary supply of nutrients. PloS One 11(1): p.e0146976. https://doi.org/10.1371/journal.pone.0146976
}
\usage{
genus_food_2011
}
\description{
Edible food supplies across 225 individual foods (grams/person/day), separated by age and sex group.
}
\details{
Countries marked with a "*" have insufficient data to estimate nutrient supplies. Each file includes a median and 95 percent uncertainty intervals for each year, calculated from 1,000 monte carlo simulations of potential nutrients supplied by the foods available in each country.
}
\keyword{datasets}
